requirements.txt
matplotlib
numpy
tensorflow
opencv-python-headless
streamlit
PILLOW
scipy
